Install go-jwt & Implement signup, login API

+++ src/entity/model/project_user.go
@@ -1,8 +1,8 @@
 package model
 
 type ProjectUser struct {
-	ID        int    `json:"id"`
-	UserID    int    `json:"user_id"`
-	ProjectID int    `json:"project_id"`
+	ID        int64  `json:"id"`
+	UserID    int64  `json:"user_id"`
+	ProjectID int64  `json:"project_id"`
 	Role      string `json:"role"` // admin, write, read
 }
